Abstract endpoint tests
